requirements.txt
streamlit
python-docx
crewai
python-dotenv
openai